[cmd/opampsupervisor] Fix supervisor ignoring RemoteConfig messages (#42497)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
The supervisor will sometimes ignore certain Remote Config messages from
the server. Specifically if the supervisor does not think the config is
different. Instead it should still respond. In the event the supervisor
does not think the config is different, it will send an applied status
(functionally the config that was sent is being ran, so applied is
appropriate here).

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
Fixes #42474

<!--Describe what testing was performed and which tests were added.-->
Unit tests updated

<!--Describe the documentation added.-->

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Co-authored-by: Evan Bradley <11745660+evan-bradley@users.noreply.github.com>